requirements.txt
xmi-reader